sandbox/requirements.txt
Django>=3.2
wagtail>=5.2
django-debug-toolbar==3.2.2
-e .[docs,test]
